requirements.txt
django
pillow 
gunicorn
python-decouple
psycopg[binary]
dj-database-url
requests
whitenoise
dotenv
django-allauth[socialaccount]
django-allauth-ui
django-widget-tweaks
slippers
stripe